+++ internal/gomod/package.go
@@ -108,7 +108,7 @@ func LoadModulesFromPackages(moduleDir, packagePattern string) ([]Module, error)
 		return nil, fmt.Errorf("failed to parse `go list` output: %w", err)
 	}
 
-	modules, err := convertPackages(moduleDir, pkgMap)
+	modules, err := convertPackagesToModules(moduleDir, pkgMap)
 	if err != nil {
 		return nil, fmt.Errorf("failed to convert packages to modules: %w", err)
 	}
@@ -139,50 +139,64 @@ func parsePackages(reader io.Reader) (map[string][]Package, error) {
 			return nil, err
 		}
 
-		if pkg.Standard {
-			log.Debug().
-				Str("package", pkg.ImportPath).
-				Str("reason", "part of standard library").
-				Msg("skipping package")
-			continue
+		if pkg.Error != nil {
+			return nil, fmt.Errorf("failed to load package: %w", pkg.Error)
 		}
-		if pkg.Module == nil {
+
+		var coordinates string
+		if pkg.Standard {
+			coordinates = StdlibModulePath
+		} else if pkg.Module == nil {
 			log.Debug().
 				Str("package", pkg.ImportPath).
 				Str("reason", "no associated module").
 				Msg("skipping package")
 			continue
+		} else {
+			coordinates = pkg.Module.Coordinates()
 		}
 
-		pkgs, ok := pkgsMap[pkg.Module.Coordinates()]
+		pkgs, ok := pkgsMap[coordinates]
 		if !ok {
-			pkgsMap[pkg.Module.Coordinates()] = []Package{pkg}
+			pkgsMap[coordinates] = []Package{pkg}
 		} else {
-			pkgsMap[pkg.Module.Coordinates()] = append(pkgs, pkg)
+			pkgsMap[coordinates] = append(pkgs, pkg)
 		}
 	}
 
 	return pkgsMap, nil
 }
 
-func convertPackages(mainModuleDir string, pkgsMap map[string][]Package) ([]Module, error) {
+func convertPackagesToModules(mainModuleDir string, pkgsMap map[string][]Package) ([]Module, error) {
 	modules := make([]Module, 0, len(pkgsMap))
-	vendoring := IsVendoring(mainModuleDir)
+	isVendoring := IsVendoring(mainModuleDir)
 
-	for _, pkgs := range pkgsMap {
-		var module *Module
-		if len(pkgs) > 0 {
-			module = pkgs[0].Module
-			if module == nil {
-				// Shouldn't ever happen, because packages without module are not collected to pkgsMap.
-				// We do the nil check anyway to make linters happy. :)
-				return nil, fmt.Errorf("no module is associated with package %s", pkgs[0].ImportPath)
-			}
-		} else {
+	for coordinates, pkgs := range pkgsMap {
+		if len(pkgs) == 0 {
 			continue
 		}
 
-		if !module.Main && vendoring {
+		var (
+			module *Module
+			err    error
+		)
+
+		if coordinates == StdlibModulePath {
+			module, err = LoadStdlibModule()
+			if err != nil {
+				return nil, fmt.Errorf("failed to load stdlib module: %w", err)
+			}
+		} else {
+			module = pkgs[0].Module
+		}
+
+		if module == nil {
+			// Shouldn't ever happen, because packages without module are not collected to pkgsMap.
+			// We do the nil check anyway to make linters happy. :)
+			return nil, fmt.Errorf("no module is associated with package %s", pkgs[0].ImportPath)
+		}
+
+		if !module.Main && isVendoring {
 			module.Vendored = true
 			vendorPath := filepath.Join(mainModuleDir, "vendor", module.Path)
 
@@ -194,66 +208,33 @@ func convertPackages(mainModuleDir string, pkgsMap map[string][]Package) ([]Modu
 			}
 		}
 
-		// In case of replacement, the replaced module
-		// doesn't have a dir, but the replacement does.
-		// We need the module dir to be able to construct
-		// filepaths relative to the module.
-		moduleDir := module.Dir
-		if moduleDir == "" && module.Replace != nil {
-			moduleDir = module.Replace.Dir
-		}
-
-		for i := range pkgs {
-			var pkgFiles []string
-			pkgFiles = append(pkgFiles, pkgs[i].GoFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].CgoFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].CFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].CXXFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].MFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].HFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].FFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].SFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].SwigFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].SwigCXXFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].SysoFiles...)
-			pkgFiles = append(pkgFiles, pkgs[i].EmbedFiles...)
-
-			for _, file := range pkgFiles {
-				filePath, err := makePackageFileRelativeToModule(moduleDir, pkgs[i].Dir, file)
-				if err == nil {
-					if module.Replace != nil {
-						module.Replace.Files = append(module.Replace.Files, filePath)
-					} else {
-						module.Files = append(module.Files, filePath)
-					}
-				} else {
-					return nil, err
-				}
-			}
-
-			sort.Slice(module.Files, func(i, j int) bool {
-				return module.Files[i] < module.Files[j]
-			})
-		}
-
 		modules = append(modules, *module)
+	}
+
+	for i := range modules {
+		var pkgs []Package
+		if modules[i].Path == StdlibModulePath {
+			pkgs = pkgsMap[StdlibModulePath]
+		} else {
+			pkgs = pkgsMap[modules[i].Coordinates()]
+		}
+
+		for j := range pkgs {
+			pkgs[j].Module = nil // we don't need this anymore
+			modules[i].Packages = append(modules[i].Packages, pkgs[j])
+		}
+
+		sortPackages(modules[i].Packages)
 	}
 
 	return modules, nil
 }
 
-func makePackageFileRelativeToModule(moduleDir, pkgDir, file string) (string, error) {
-	moduleDirAbs, err := filepath.Abs(moduleDir)
-	if err != nil {
-		return "", err
-	}
-
-	fp, err := filepath.Rel(moduleDirAbs, filepath.Join(pkgDir, file))
-	if err != nil {
-		return "", err
-	}
-
-	return strings.ReplaceAll(fp, "\\", "/"), nil
+// sortPackages sorts a given Package slice ascending by import path.
+func sortPackages(pkgs []Package) {
+	sort.Slice(pkgs, func(i, j int) bool {
+		return pkgs[i].ImportPath < pkgs[j].ImportPath
+	})
 }
 
 // toRelativePackagePath ensures that Go will interpret the given packagePattern