+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## v1.1.0 (unreleased)
+
+### Enhancements
+
+* Add option to assert detected licenses ([#96](https://github.com/CycloneDX/cyclonedx-gomod/pull/97) via [#97](https://github.com/CycloneDX/cyclonedx-gomod/pull/97))
+  * This will move licenses from `evidence/licenses` to `licenses`, which helps with SBOM ingestion in some cases
+* `app`: Add option to include packages in application SBOM ([#85](https://github.com/CycloneDX/cyclonedx-gomod/issues/85) via [#92](https://github.com/CycloneDX/cyclonedx-gomod/pull/92))
+* `app`: The `-packages` and `-files` options are now also applied to the standard library component (when `-std` is used) ([#84](https://github.com/CycloneDX/cyclonedx-gomod/issues/84) via [#92](https://github.com/CycloneDX/cyclonedx-gomod/pull/92))
+  * Thanks [TheDiveO](https://github.com/TheDiveO) for reporting!
+* Package URLs now include a `type` qualifier to better differentiate between modules and packages (via [`1c4b136`](https://github.com/CycloneDX/cyclonedx-gomod/pull/92/commits/1c4b1366ac23e6a4387f1e6d4b35b67930184aed))
+
+### Breaking Changes
+
+* `app`: `-files` can now only be used in conjunction with `-packages`
+* `app`: Files are now represented as subcomponents of packages
+
+### Miscellaneous
+
+* The `go` prefix is no longer stripped from Go versions
+  * e.g. the standard library module will now appear as `pkg:golang/std@go1.17.3` instead of `pkg:golang/std@1.17.3`
+
+### Dependency Updates
+
+* Update `github.com/rs/zerolog` from `v1.25.0` to `v1.26.0`
+
+### Building and Packaging
+
+* Bump `golang` container base images from `1.17.2` to `1.17.3` (via [#95](https://github.com/CycloneDX/cyclonedx-gomod/pull/95))
+* Reference container base images by their SHA digest ([#89](https://github.com/CycloneDX/cyclonedx-gomod/issues/89) via [#90](https://github.com/CycloneDX/cyclonedx-gomod/pull/90))
+* Introduce multi-platform container image builds ([#87](https://github.com/CycloneDX/cyclonedx-gomod/issues/87) via [#90](https://github.com/CycloneDX/cyclonedx-gomod/pull/90))
+
 ## v1.0.0
 
 ### Enhancements